updated authoriztion service, and bug fixes

+++ server/routes/api/project-routes.js
@@ -1,22 +1,10 @@
 const router = require('express').Router();
 const connection = require('../../config/connection');
 const dbQuery = require('../../utils/dbQuery');
+const { checkProjectAccess, checkProjectPrivilege } = require('../../utils/authorize');
 
 router.get('/', async (req, res) => {
     if(req.session.user) {
-        // find all user projects
-        // try {
-        //   const [rows] = await connection.query(
-        //     `SELECT * FROM user_project
-        //     INNER JOIN project ON user_project.project_id = project.id
-        //     WHERE user_project.user_id = (SELECT user.id FROM user WHERE user.user_name = '${req.session.user}');`);
-
-        //   res.status(200).json(rows);
-        // } catch (error) {
-        //   console.error("Error executing queries:", error);
-        //   res.status(500).send("An error occurred while executing the database queries.");
-        // }
-
         // find all user projects
         const query = `SELECT * FROM user_project
             INNER JOIN project ON user_project.project_id = project.id
@@ -25,38 +13,32 @@ router.get('/', async (req, res) => {
         const result = await dbQuery('GET', { query });
         res.status(200).json(result);
     } else {
-        console.log('No user');
         res.send(null);
     }
 });
 
 router.get('/:id', async (req, res) => {
     // find all tasks for project by project `id`
+
     if(req.session.user) {
-        // try {
-        //   const [rows] = await connection.query(
-        //     `SELECT task.id, task_name, task_description, project_id, assigned_to_id, user_name AS assigned_to, priority_id, level, status_id, status FROM task
-        //     INNER JOIN user ON task.assigned_to_id = user.id
-        //     INNER JOIN priority ON task.priority_id = priority.id
-        //     INNER JOIN status ON task.status_id = status.id
-        //     WHERE task.project_id = ${req.params.id};`);
 
-        //   res.status(200).json(rows);
-        // } catch (error) {
-        //   console.error("Error executing queries:", error);
-        //   res.status(500).send("An error occurred while executing the database queries.");
-        // }
+        // check authorization (session-user to project id)
+        const auth = await checkProjectAccess({project_id: req.params.id, 
+            user_name: req.session.user});
+        
+        if (auth.authorized == true) {
+            const query = `SELECT task.id, task_name, task_description, project_id, assigned_to_id, user_name AS assigned_to, priority_id, level, status_id, status FROM task
+                INNER JOIN user ON task.assigned_to_id = user.id
+                INNER JOIN priority ON task.priority_id = priority.id
+                INNER JOIN status ON task.status_id = status.id
+                WHERE task.project_id = ${req.params.id};`;
 
-        const query = `SELECT task.id, task_name, task_description, project_id, assigned_to_id, user_name AS assigned_to, priority_id, level, status_id, status FROM task
-            INNER JOIN user ON task.assigned_to_id = user.id
-            INNER JOIN priority ON task.priority_id = priority.id
-            INNER JOIN status ON task.status_id = status.id
-            WHERE task.project_id = ${req.params.id};`;
-
-        const result = await dbQuery('GET', { query });
-        res.status(200).json(result);
+            const result = await dbQuery('GET', { query });
+            res.status(200).json(result);
+        } else {
+            res.status(200).send(null);
+        }
     } else {
-        console.log('No user');
         res.send(null);
     }
 });
@@ -64,26 +46,6 @@ router.get('/:id', async (req, res) => {
 router.post('/', async (req, res) => {
   // Create new project
   if(req.session.user) {
-      // try {
-      //   const [rows] = await connection.query(
-      //     `INSERT INTO project ( project_name )
-      //     VALUES ( '${req.body.project_name}' );`);
-                
-      //   try {
-      //     const [data] = await connection.query(
-      //       `INSERT INTO user_project ( user_id, project_id, permission_id )
-      //       VALUES ( (SELECT id FROM user WHERE user_name = '${req.session.user}' ), ${rows.insertId}, 3 );`);
-
-      //     res.status(200).json(rows);
-      //   } catch (error) {
-      //     console.error("Error executing queries:", error);
-      //     res.status(500).send("An error occurred while executing the database queries.");
-      //   }
-      // } catch (error) {
-      //   console.error("Error executing queries:", error);
-      //   res.status(500).send("An error occurred while executing the database queries.");
-      // }
-
       let query = `INSERT INTO project ( project_name )
           VALUES ( '${req.body.project_name}' );`;
 
@@ -96,57 +58,52 @@ router.post('/', async (req, res) => {
 
       res.status(200).json(result2);
   } else {
-      console.log('No user');
       res.send(null);
   }
 });
 
 router.put('/:id', async (req, res) => {
     // update project by project `id`
+
     if(req.session.user) {
-        // try {
-        //   const [rows] = await connection.query(
-        //     `UPDATE project
-        //     SET project_name = '${req.body.project_name}'
-        //     WHERE id = ${req.params.id};`);
+        // check authorization (session-user to project permission 3)
+        const auth = await checkProjectPrivilege({project_id: req.params.id, 
+                user_name: req.session.user,
+                permission_id: 3});
+        
+        if (auth.authorized == true) {
+            const query = `UPDATE project
+                SET project_name = '${req.body.project_name}'
+                WHERE id = ${req.params.id};`;
 
-        //   res.status(200).json(rows);
-        // } catch (error) {
-        //   console.error("Error executing queries:", error);
-        //   res.status(500).send("An error occurred while executing the database queries.");
-        // }
-
-        const query = `UPDATE project
-            SET project_name = '${req.body.project_name}'
-            WHERE id = ${req.params.id};`;
-
-        const result = await dbQuery('PUT', { query });
-        res.status(200).json(result);
+            const result = await dbQuery('PUT', { query });
+            res.status(200).json(result);
+        } else {
+            res.status(200).send(null);
+        }
     } else {
-        console.log('No user');
         res.send(null);
     }
 });
 
 router.delete('/:id', async (req, res) => {
     // delete project by project `id`
+
     if(req.session.user) {
-        // try {
-        //   const [rows] = await connection.query(
-        //     `DELETE FROM project WHERE id = ${req.params.id};`);
+        // check authorization (session-user to project permission 3)
+        const auth = await checkProjectPrivilege({project_id: req.params.id, 
+                user_name: req.session.user,
+                permission_id: 3});
+        
+        if (auth.authorized == true) {
+            const query = `DELETE FROM project WHERE id = ${req.params.id};`;
 
-        //   res.status(200).json(rows);
-        // } catch (error) {
-        //   console.error("Error executing queries:", error);
-        //   res.status(500).send("An error occurred while executing the database queries.");
-        // }
-
-        const query = `DELETE FROM project WHERE id = ${req.params.id};`;
-
-        const result = await dbQuery('DELETE', { query });
-        res.status(200).json(result);
+            const result = await dbQuery('DELETE', { query });
+            res.status(200).json(result);
+        } else {
+            res.status(200).send(null);
+        }
     } else {
-        console.log('No user');
         res.send(null);
     }
 });
